requirements.txt
decorator==4.2.1
Django==1.11.9
django-filters==0.2.1
django-rest-framework==0.1.0
djangorestframework==3.7.7
psycopg2==2.7.4
public==0.0.65
pytz==2018.3
self==0.0.27
